requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# ── Data validation (FastAPI uses pydantic v2) ────────────────────────────────
pydantic>=2.0.0

# ── Data / numerics ───────────────────────────────────────────────────────────
pandas>=1.4.0
numpy>=1.22.0
numpy-financial>=1.0.0

# ── Excel / CSV parsing ───────────────────────────────────────────────────────
openpyxl>=3.0.0          # .xlsx read/write (used by fiscal impact rates)
xlrd>=2.0.0              # legacy .xls support
python-calamine>=0.1.0   # fast Rust-backed Excel reader (optional speedup)

# ── HTTP / scraping ───────────────────────────────────────────────────────────
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# ── Database (accounting/incentives engine) ───────────────────────────────────
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dateutil>=2.8.0
python-dotenv>=0.19.0
python-multipart>=0.0.9      # required by FastAPI for form data
anthropic>=0.20.0            # Claude API SDK (memo endpoint fallback)
